+++ internal/resolvers/youtube/resolver.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"net/http"
 	"net/url"
+	"regexp"
 	"time"
 
 	"github.com/Chatterino/api/pkg/cache"
@@ -19,7 +20,7 @@ import (
 )
 
 const (
-	youtubeTooltip = `<div style="text-align: left;">
+	youtubeVideoTooltip = `<div style="text-align: left;">
 <b>{{.Title}}</b>
 <br><b>Channel:</b> {{.ChannelTitle}}
 <br><b>Duration:</b> {{.Duration}}
@@ -28,14 +29,29 @@ const (
 <br><span style="color: #2ecc71;">{{.LikeCount}} likes</span>&nbsp;•&nbsp;<span style="color: #e74c3c;">{{.DislikeCount}} dislikes</span>
 </div>
 `
+
+	youtubeChannelTooltip = `<div style="text-align: left;">
+<b>{{.Title}}</b>
+<br><b>Joined Date:</b> {{.JoinedDate}}
+<br><b>Subscribers:</b> {{.Subscribers}}
+<br><b>Views:</b> {{.Views}}
+</div>
+`
 )
 
 var (
-	videoCache = cache.New("youtube_videos", load, 24*time.Hour)
+	// YouTube videos are cached by video ID
+	videoCache = cache.New("youtube_videos", loadVideos, 24*time.Hour)
+	// Channels are cached by <channel_type>:<channel_id>
+	// See channelCacheKey.go for more information
+	channelCache = cache.New("youtube_channels", loadChannels, 24*time.Hour)
 
 	youtubeClient *youtubeAPI.Service
 
-	youtubeTooltipTemplate = template.Must(template.New("youtubeTooltip").Parse(youtubeTooltip))
+	youtubeVideoTooltipTemplate   = template.Must(template.New("youtubeVideoTooltip").Parse(youtubeVideoTooltip))
+	youtubeChannelTooltipTemplate = template.Must(template.New("youtubeChannelTooltip").Parse(youtubeChannelTooltip))
+
+	youtubeChannelRegex = regexp.MustCompile(`/(user|c(hannel)?)/[\w._\-']+`)
 )
 
 func New(cfg config.APIConfig) (resolvers []resolver.CustomURLManager) {
@@ -51,6 +67,26 @@ func New(cfg config.APIConfig) (resolvers []resolver.CustomURLManager) {
 		return
 	}
 
+	// Handle YouTube channels (youtube.com/c/chan, youtube.com/chan, youtube.com/user/chan)
+	resolvers = append(resolvers, resolver.CustomURLManager{
+		Check: func(url *url.URL) bool {
+			matches := youtubeChannelRegex.MatchString(url.Path)
+			return utils.IsSubdomainOf(url, "youtube.com") && matches
+		},
+		Run: func(url *url.URL, r *http.Request) ([]byte, error) {
+			channelID := getYoutubeChannelIDFromURL(url)
+
+			if channelID.chanType == InvalidChannel {
+				return resolver.NoLinkInfoFound, nil
+			}
+
+			channelCacheKey := constructCacheKeyFromChannelID(channelID)
+			apiResponse := channelCache.Get(channelCacheKey, r)
+			return json.Marshal(apiResponse)
+		},
+	})
+
+	// Handle YouTube video URLs
 	resolvers = append(resolvers, resolver.CustomURLManager{
 		Check: func(url *url.URL) bool {
 			return utils.IsSubdomainOf(url, "youtube.com")
@@ -67,6 +103,7 @@ func New(cfg config.APIConfig) (resolvers []resolver.CustomURLManager) {
 		},
 	})
 
+	// Handle shortened YouTube video URLs
 	resolvers = append(resolvers, resolver.CustomURLManager{
 		Check: func(url *url.URL) bool {
 			return url.Host == "youtu.be"